feat: add constants data for modes
